requirements.txt
# Core Django and DRF
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0

# Database
psycopg[binary]>=3.1.0

# Configuration
python-decouple>=3.8

# API Documentation
drf-spectacular>=0.27.0

# API Extensions
django-filter>=23.0
django-cors-headers>=4.3.0

# Media handling
Pillow>=10.0.0

# Development and Testing
pytest>=7.4.0
pytest-django>=4.7.0
pytest-cov>=4.1.0
django-debug-toolbar>=4.2.0

# Code Quality
black>=23.0.0
isort>=5.12.0
ruff>=0.1.0
mypy>=1.7.0
pre-commit>=3.5.0

# Production
gunicorn>=21.2.0
redis>=5.0.0

# Optional: For advanced features
celery>=5.3.0
django-extensions>=3.2.0
